refactor : 1차 피드백 반영
1. cri -> criteria
2. repository에 count 추가
3. service에서 작성자 일치여부 검증

+++ src/main/java/com/codessquad/qna/service/AnswerService.java
@@ -4,6 +4,7 @@ import com.codessquad.qna.domain.answer.Answer;
 import com.codessquad.qna.domain.answer.AnswerRepository;
 import com.codessquad.qna.domain.question.Question;
 import com.codessquad.qna.domain.question.QuestionRepository;
+import com.codessquad.qna.domain.user.User;
 import com.codessquad.qna.exception.NotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -21,9 +22,10 @@ public class AnswerService {
         this.questionRepository = questionRepository;
     }
 
-    public Answer create(Long questionId, Answer answer) {
+    public Answer create(Long questionId, Answer answer, User user) {
         Question question = questionRepository.findById(questionId)
-                .orElseThrow(() -> new NotFoundException("해당 질문이 없습니다. id = " + questionId));
+                .orElseThrow(() -> new IllegalStateException("해당 질문이 없습니다. id = " + questionId));
+        answer.setWriter(user);
         answer.setQuestion(question);
         question.addAnswer(answer);
         return answerRepository.save(answer);
@@ -50,10 +52,14 @@ public class AnswerService {
     }
 
     @Transactional
-    public Long deleteById(Long id) {
+    public Long deleteById(Long id, User user) {
         Answer answer = answerRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException("해당 답변이 없습니다. id = " + id));
+                .orElseThrow(() -> new IllegalStateException("해당 답변이 없습니다. id = " + id));
+        if (!answer.isWrittenBy(user)) {
+            throw new IllegalStateException("자신이 작성한 답변만 삭제할 수 있습니다.");
+        }
         Question question = answer.getQuestion();
+                .orElseThrow(() -> new NotFoundException("해당 답변이 없습니다. id = " + id));
         answer.delete();
         question.downCountOfAnswer();
